Clean up and canonicalize the command invocation section.
This patch describes the `help` command and also changes the
typography used in the command invocation section. The typography is
altered to look more like a traditional man-page, where each flag is
listed apart from the others. This makes for a more readable
overview.

Also use some horizontal rulers to separate the different command
invocation types, further making it easier to read.