Go through the README.md file and clean up.
Minor wording touch-ups here and there. Moved things around and
improved the typography some more. I hope it is better and easier to
read. I decided to remove some typography in order to hopefully make
the file less daunting on the eye. Also some more spread was added
around lists, so they are more split. Finally, headings of size 4 were
made into headings of size 3 so the structure is following the
document better.