Fixed a bug when build or install ports normal EXIT signals were not flushed
In some cases it led to behaviour that required tracing the whole agner_fetch fsm. The problem
was that gen_fsm was receiving an EXIT signal within gen_fsm2 loop, which runs handle_state every time
it loops. That resulted in handle_state(buildable, State) being invoked twice and therefore a 'next' event
that signals state transition was sent twice and the second one was immediately processed by installable/2
which made it proceed to the next state immediately. For installable/2, this next step is a routine fsm shutdown.
Long story short, it make agner_fetch skip the installation part.